+++ packages/genesys-spark-components/scripts/generate-gux-icon-types-file.js
@@ -1,19 +1,36 @@
 #! /usr/bin/env node
 
 const fs = require('fs');
+const path = require('path');
 const prettier = require('prettier');
 
-const svgExtentionRegex = new RegExp('.svg$', 'i');
+const svgExtensionRegex = new RegExp('.svg$', 'i');
+let variants = [];
 
-const variants = fs
-  .readdirSync(__dirname + '/../src/components/stable/gux-icon/icons')
-  .filter(filename => filename.match(svgExtentionRegex))
-  .map(filename => filename.replace(svgExtentionRegex, ''))
-  .map(filename => `'${filename}'`)
-  .sort()
-  .join(' | ');
+const getAllVariants = directory => {
+  const fileList = fs.readdirSync(directory);
+  fileList.forEach(file => {
+    file = path.join(directory, file);
+    const fileStatus = fs.statSync(file);
+    if (fileStatus.isDirectory() && !file.includes('legacy')) {
+      // Recursively search sub directories excluding legacy.
+      variants = variants.concat(getAllVariants(file));
+    } else {
+      variants.push(file.split('icons/').pop());
+    }
+  });
 
-const content = `export type GuxIconIconName = ${variants};`;
+  return variants
+    .filter(filename => filename.match(svgExtensionRegex))
+    .map(filename => filename.replace(svgExtensionRegex, ''))
+    .map(filename => `'${filename}'`)
+    .sort()
+    .join(' | ');
+};
+
+const content = `export type GuxIconIconName = ${getAllVariants(
+  __dirname + '/../src/components/stable/gux-icon/icons'
+)}`;
 
 const filePath =
   __dirname + '/../src/components/stable/gux-icon/gux-icon.types.ts';