+++ packages/genesys-spark-components/src/components/beta/gux-chart-donut/gux-chart-donut.tsx
@@ -131,8 +131,8 @@ export class GuxDonutChart {
   parseData() {
     if (!this.outerRadius && !this.innerRadius) {
       logError(
-        'gux-chart-donut',
-        '[gux-chart-donut] requires at least one of outer-radius or inner-radius'
+        this.root,
+        'requires at least one of outer-radius or inner-radius'
       );
     }