+++ packages/genesys-spark-components/src/components/beta/gux-table-toolbar/gux-table-toolbar-action/gux-table-toolbar-action.tsx
@@ -8,6 +8,7 @@ import translationResources from './i18n/en.json';
 import { GuxTableToolbarActionAccent } from '../gux-table-toolbar-action-accents.types';
 
 @Component({
+  styleUrl: 'gux-table-toolbar-action.less',
   tag: 'gux-table-toolbar-action',
   shadow: true
 })