+++ packages/genesys-spark-components/src/components/stable/gux-form-field/components/gux-form-field-number/gux-form-field-number.tsx
@@ -33,7 +33,8 @@ import {
   clearInput,
   getComputedLabelPosition,
   hasContent,
-  validateFormIds
+  validateFormIds,
+  getSlottedInput
 } from '../../gux-form-field.service';
 import { trackComponent } from '@utils/tracking/usage';
 import componentResources from './i18n/en.json';
@@ -48,7 +49,7 @@ import componentResources from './i18n/en.json';
 @Component({
   styleUrl: 'gux-form-field-number.less',
   tag: 'gux-form-field-number',
-  shadow: true
+  shadow: { delegatesFocus: true }
 })
 export class GuxFormFieldNumber {
   private getI18nValue: GetI18nValue;
@@ -105,7 +106,6 @@ export class GuxFormFieldNumber {
       this.root,
       componentResources as ILocalizedComponentResources
     );
-
     this.setInput();
     this.setLabel();
 
@@ -184,7 +184,10 @@ export class GuxFormFieldNumber {
   }
 
   private setInput(): void {
-    this.input = this.root.querySelector('input[type="number"][slot="input"]');
+    this.input = getSlottedInput(
+      this.root,
+      'input[type="number"][slot="input"]'
+    );
 
     this.hasContent = hasContent(this.input);