+++ packages/genesys-spark-components/src/components/stable/gux-form-field/components/gux-form-field-range/gux-form-field-range.tsx
@@ -26,7 +26,8 @@ import {
 import { GuxFormFieldLabelPosition } from '../../gux-form-field.types';
 import {
   getComputedLabelPosition,
-  validateFormIds
+  validateFormIds,
+  getSlottedInput
 } from '../../gux-form-field.service';
 import { trackComponent } from '@utils/tracking/usage';
 
@@ -171,7 +172,10 @@ export class GuxFormField {
   }
 
   private setInput(): void {
-    this.input = this.root.querySelector('input[type="range"][slot="input"]');
+    this.input = getSlottedInput(
+      this.root,
+      'input[type="range"][slot="input"]'
+    );
 
     preventBrowserValidationStyling(this.input);