+++ web-apps/genesys-spark-examples/src/utils/generateLegacyIcons.js
@@ -9,14 +9,23 @@ function generateHTML(path) {
     if (!icon.endsWith('.svg')) continue;
 
     const iconName = icon.substring(0, icon.length - 4); // remove .svg suffix
+    let mappedIconName = iconName;
 
-    output += `<div class="icon-example">
-  <gux-icon icon-name="legacy/${iconName}" class="example" decorative="true"></gux-icon>
-  <div class="icon-name">${iconName} (original)</div>
+    // if a legacy icon is added and there is already an icon in the legacy folder with the same name,
+    // the newly added icon should be given a suffix following the pattern: "-alt-1"
+    if (/-alt-\d$/.test(iconName)) {
+      mappedIconName = iconName.substring(0, iconName.length - 6); //remove -alt-1 suffix
+    }
+
+    output += `<div class="icon-example-group">
+  <div class="icon-example">
+    <gux-icon icon-name="legacy/${iconName}" class="example" decorative="true"></gux-icon>
+    <div class="icon-name">legacy/${iconName}</div>
   </div>
-<div class="icon-example">
-  <gux-icon icon-name="${iconName}" class="example" decorative="true"></gux-icon>
-  <div class="icon-name">${iconName} (mapped)</div>
+  <div class="icon-example">
+    <gux-icon icon-name="${mappedIconName}" class="example" decorative="true"></gux-icon>
+    <div class="icon-name">${mappedIconName} (mapped)</div>
+  </div>
 </div>
 `;
   }